(PE-27344) Add update_effective_date function for updating EFFECTIVE_DATE in shim
This creates a function that takes an effective date string, a pe-installer-shim branch, and a ticket identifier, and will update the EFFECTIVE_DATE variable in the shim.  It verifies that the date given is in the correct format, clones and checks out the given branch of pe-installer-shim, modifies the EFFECTIVE_DATE variable, and pushes a commit to origin with that change.