(maint) Update release branch script to use codename
Instead of using a `branch_from` parameter, this commit updates the
script to instead use a codename. This will hopefully account for the
difference in using versions or codenames as branch names across our
different repos.

The way this will work is that the script will try to check out a branch
by first using the codename, and then will use the version parsing logic
to check out a version branch if the codename checkout fails. This is
not the most elegant solution ever, but should work